# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Firewall LLM Unified API
  version: 0.1.0
  description: >
    OpenAI-compatible unified entry point. One contract for all provider adapters
    (Python and Rust gateway branches).
servers:
  - url: http://localhost:8080
paths:
  /v1/chat/completions:
    post:
      summary: Create a chat completion
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCompletionRequest"
      responses:
        "200":
          description: Completion result (single object or SSE stream)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatCompletionResponse"
            text/event-stream:
              schema:
                $ref: "#/components/schemas/ChatCompletionChunk"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Blocked"
        "422":
          $ref: "#/components/responses/InvalidRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
        "502":
          $ref: "#/components/responses/UpstreamError"
  /healthz:
    get:
      summary: Liveness probe
      responses:
        "200":
          description: Service is up
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [ok]
  /metrics:
    get:
      summary: Prometheus metrics
      description: >
        Protected endpoint (R15). Accepts admin tokens or least-privilege
        metrics tokens (FWLLM_METRICS_TOKENS, /metrics only); ordinary client
        tokens are rejected. Prometheus scrapes it with bearer_token_file.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Metrics in Prometheus text format
          content:
            text/plain:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Blocked"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    Unauthorized:
      description: Missing or invalid client API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Blocked:
      description: Request blocked by inspector (injection/DLP) or policy
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InvalidRequest:
      description: Malformed request body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Quota or rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UpstreamError:
      description: Provider adapter failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [system, user, assistant, tool]
        content:
          type: [string, "null"]
        name:
          type: string
        tool_calls:
          type: array
          items:
            type: object
    ChatCompletionRequest:
      type: object
      required: [model, messages]
      properties:
        model:
          type: string
          description: Logical model name; router maps it to a concrete provider model
        messages:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ChatMessage"
        stream:
          type: boolean
          default: false
        temperature:
          type: number
          minimum: 0
          maximum: 2
        top_p:
          type: number
          minimum: 0
          maximum: 1
        max_tokens:
          type: integer
          minimum: 1
        stop:
          type: [string, array]
          items:
            type: string
        metadata:
          type: object
          additionalProperties: true
          description: Client metadata (user id, project, etc.) used for metering/policies
    Usage:
      type: object
      required: [prompt_tokens, completion_tokens, total_tokens]
      properties:
        prompt_tokens:
          type: integer
        completion_tokens:
          type: integer
        total_tokens:
          type: integer
    Choice:
      type: object
      required: [index, message]
      properties:
        index:
          type: integer
        message:
          $ref: "#/components/schemas/ChatMessage"
        finish_reason:
          type: [string, "null"]
          enum: [stop, length, tool_calls, content_filter, null]
    ChatCompletionResponse:
      type: object
      required: [id, object, created, model, choices]
      properties:
        id:
          type: string
        object:
          type: string
          enum: [chat.completion]
        created:
          type: integer
        model:
          type: string
          description: Provider model that actually served the request
        routed_from:
          type: string
          description: Original logical model from the request (present if remapped)
        choices:
          type: array
          items:
            $ref: "#/components/schemas/Choice"
        usage:
          $ref: "#/components/schemas/Usage"
    ChatCompletionChunk:
      type: object
      required: [id, object, created, model, choices]
      properties:
        id:
          type: string
        object:
          type: string
          enum: [chat.completion.chunk]
        created:
          type: integer
        model:
          type: string
        choices:
          type: array
          items:
            type: object
            properties:
              index:
                type: integer
              delta:
                type: object
                properties:
                  role:
                    type: string
                  content:
                    type: string
              finish_reason:
                type: [string, "null"]
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message, type]
          properties:
            message:
              type: string
            type:
              type: string
              enum:
                - invalid_request_error
                - authentication_error
                - permission_error   # blocked by inspector/policy
                - rate_limit_error
                - upstream_error
            code:
              type: string
            details:
              type: object
              additionalProperties: true